adds screens to the example
