+++ example/android/app/src/main/java/com/revenuecat/purchases_sample/MainApplication.java
@@ -3,6 +3,7 @@ package com.revenuecat.purchases_sample;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 import com.reactlibrary.RNPurchasesPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
@@ -24,6 +25,7 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
+            new RNGestureHandlerPackage(),
             new RNPurchasesPackage()
       );
     }
