+++ example/android/settings.gradle
@@ -1,4 +1,6 @@
 rootProject.name = 'ReactNativeSample'
+include ':react-native-gesture-handler'
+project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
 include ':react-native-purchases'
 project(':react-native-purchases').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-purchases/android')
