+++ example/package.json
@@ -11,7 +11,9 @@
     "metro-config": "^0.51.1",
     "react": "16.6.3",
     "react-native": "0.57.7",
-    "react-native-purchases": "file:.."
+    "react-native-gesture-handler": "^1.0.15",
+    "react-native-purchases": "file:..",
+    "react-navigation": "^3.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-0",
@@ -19,6 +21,8 @@
     "babel-jest": "23.6.0",
     "jest": "23.6.0",
     "metro-react-native-babel-preset": "0.50.0",
+    "prop-types": "^15.7.1",
+    "react-native-screens": "^1.0.0-alpha.22",
     "react-test-renderer": "16.6.3"
   },
   "jest": {
